feat: Make the task lists work. Fix bug
Make the task lists in the right column work. Add drag icons to task lists. Make saving lists of tasks in the localStorage. Make the list open at the moment of the last work in the program. Make selection of the active list of tasks, according to the styles. Fix container size bug when rendering tasks with a long list less than 600px